Move GitHub 1197 test case under its own package
