+++ src/test/java/test_result/github1197/GitHub1197Test.java
@@ -1,4 +1,4 @@
-package test_result;
+package test_result.github1197;
 
 import org.testng.ITestNGListener;
 import org.testng.TestNG;
